requirements.txt
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
obd>=0.7.1
streamlit>=1.20.0
tensorflow==2.12.0
joblib>=1.2.0
xgboost>=1.7.0
python-dotenv>=0.19.0
